+++ crates/test-utils/src/test_builders.rs
@@ -225,6 +225,7 @@ macro_rules! build_test_by_mode {
             in_tracing_mode: $in_tracing_mode,
             libraries: ::alloc::vec::Vec::default(),
             handlers: ::alloc::vec::Vec::new(),
+            trace_handlers: ::alloc::vec::Vec::new(),
             add_modules: ::alloc::vec::Vec::default(),
         }
     }};
@@ -251,6 +252,7 @@ macro_rules! build_test_by_mode {
             in_tracing_mode: $in_tracing_mode,
             libraries: ::alloc::vec::Vec::default(),
             handlers: ::alloc::vec::Vec::new(),
+            trace_handlers: ::alloc::vec::Vec::new(),
             add_modules: ::alloc::vec::Vec::default(),
         }
     }};
@@ -282,6 +284,7 @@ macro_rules! build_test_by_mode {
             in_tracing_mode: $in_tracing_mode,
             libraries: ::alloc::vec::Vec::default(),
             handlers: ::alloc::vec::Vec::new(),
+            trace_handlers: ::alloc::vec::Vec::new(),
             add_modules: ::alloc::vec::Vec::default(),
         }
     }};
@@ -315,6 +318,7 @@ macro_rules! build_test_by_mode {
             in_tracing_mode: $in_tracing_mode,
             libraries: ::alloc::vec::Vec::default(),
             handlers: ::alloc::vec::Vec::new(),
+            trace_handlers: ::alloc::vec::Vec::new(),
             add_modules: ::alloc::vec::Vec::default(),
         }
     }};